+++ packages/eslint-plugin-svelte/src/types.ts
@@ -240,6 +240,8 @@ export interface SourceCode {
 
 	getLines(): string[];
 
+	getDeclaredVariables(node: TSESTree.Node): Variable[];
+
 	getAllComments(): AST.Comment[];
 
 	getComments(node: NodeOrToken): {