+++ packages/eslint-plugin-svelte/tests/fixtures/rules/no-navigation-without-base/invalid/goto-aliased01-input.svelte
@@ -1,0 +1,5 @@
+<script>
+	import { goto as alias } from '$app/navigation';
+
+	alias('/foo');
+</script>