+++ packages/eslint-plugin-svelte/tests/fixtures/rules/no-navigation-without-base/invalid/goto-base-not-as-prefix01-input.svelte
@@ -1,0 +1,7 @@
+<script>
+	import { base } from '$app/paths';
+	import { goto } from '$app/navigation';
+
+	goto('/foo/' + base);
+	goto(`/foo/${base}`);
+</script>