+++ packages/eslint-plugin-svelte/tests/fixtures/rules/no-navigation-without-base/valid/goto-base-namespace-import01-input.svelte
@@ -1,0 +1,8 @@
+<script>
+	import * as paths from '$app/paths';
+	import { goto } from '$app/navigation';
+
+	// eslint-disable-next-line prefer-template -- Testing both variants
+	goto(paths.base + '/foo/');
+	goto(`${paths.base}/foo/`);
+</script>