+++ packages/eslint-plugin-svelte/tests/fixtures/rules/no-navigation-without-base/valid/ignorePushState/pushState-ignored01-input.svelte
@@ -1,0 +1,5 @@
+<script>
+	import { pushState } from '$app/navigation';
+
+	pushState('/foo');
+</script>