+++ packages/eslint-plugin-svelte/tests/fixtures/rules/no-navigation-without-base/valid/link-base-aliased01-input.svelte
@@ -1,0 +1,6 @@
+<script>
+	import { base as alias } from '$app/paths';
+</script>
+
+<a href={alias + '/foo/'}>Click me!</a>;
+<a href={`${alias}/foo/`}>Click me!</a>;