+++ packages/eslint-plugin-svelte/tests/fixtures/rules/prefer-const/invalid/test01-output.svelte
@@ -1,0 +1,11 @@
+<script>
+	let { prop1, prop2 } = $props();
+	const zero = 0;
+	const state = $state(0);
+	const raw = $state.raw(0);
+	const doubled = state * 2;
+	let derived = $derived(state * 2);
+	const calculated = calc();
+	let derivedBy = $derived.by(calc());
+	let noInit;
+</script>